backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.5.3
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
graphviz==0.20.3
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
PyJWT==2.10.1
pymongo==4.15.4
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.44
tinydb==4.8.2
typing_extensions==4.15.0
Werkzeug==3.1.3
